Replace .tests.reporting with genno.tests.compat.*ixmp*
