--- ixmp/tests/data/report-describe.txt
@@ -1,110 +1,0 @@
-'a:':
-- sum(dimensions=['i'], weights=None, ...)
-- 'a:i':
-  - data_for_quantity('par', 'a', 'value', ...)
-  - 'scenario':
-    - <ixmp.core.Scenario object at {id}>
-  - 'config':
-    - {{'filters': {{}}}}
-
-'b:':
-- sum(dimensions=['j'], weights=None, ...)
-- 'b:j':
-  - data_for_quantity('par', 'b', 'value', ...)
-  - 'scenario' (above)
-  - 'config' (above)
-
-'cost-margin:':
-- data_for_quantity('equ', 'cost', 'mrg', ...)
-- 'scenario' (above)
-- 'config' (above)
-
-'cost:':
-- data_for_quantity('equ', 'cost', 'lvl', ...)
-- 'scenario' (above)
-- 'config' (above)
-
-'d:':
-- sum(dimensions=['i', 'j'], weights=None, ...)
-- 'd:i-j':
-  - data_for_quantity('par', 'd', 'value', ...)
-  - 'scenario' (above)
-  - 'config' (above)
-
-'d:i':
-- sum(dimensions=['j'], weights=None, ...)
-- 'd:i-j' (above)
-
-'d:j':
-- sum(dimensions=['i'], weights=None, ...)
-- 'd:i-j' (above)
-
-'demand-margin:j':
-- data_for_quantity('equ', 'demand', 'mrg', ...)
-- 'scenario' (above)
-- 'config' (above)
-
-'demand:':
-- sum(dimensions=['j'], weights=None, ...)
-- 'demand:j':
-  - data_for_quantity('equ', 'demand', 'lvl', ...)
-  - 'scenario' (above)
-  - 'config' (above)
-
-'f:':
-- data_for_quantity('par', 'f', 'value', ...)
-- 'scenario' (above)
-- 'config' (above)
-
-'i':
-- ['seattle', 'san-diego']
-
-'j':
-- ['new-york', 'chicago', 'topeka']
-
-'supply-margin:i':
-- data_for_quantity('equ', 'supply', 'mrg', ...)
-- 'scenario' (above)
-- 'config' (above)
-
-'supply:':
-- sum(dimensions=['i'], weights=None, ...)
-- 'supply:i':
-  - data_for_quantity('equ', 'supply', 'lvl', ...)
-  - 'scenario' (above)
-  - 'config' (above)
-
-'x:':
-- sum(dimensions=['i', 'j'], weights=None, ...)
-- 'x:i-j':
-  - data_for_quantity('var', 'x', 'lvl', ...)
-  - 'scenario' (above)
-  - 'config' (above)
-
-'x:i':
-- sum(dimensions=['j'], weights=None, ...)
-- 'x:i-j' (above)
-
-'x:j':
-- sum(dimensions=['i'], weights=None, ...)
-- 'x:i-j' (above)
-
-'z:':
-- data_for_quantity('var', 'z', 'lvl', ...)
-- 'scenario' (above)
-- 'config' (above)
-
-'all':
-- list of:
-  - 'a:i' (above)
-  - 'b:j' (above)
-  - 'cost-margin:' (above)
-  - 'cost:' (above)
-  - 'd:i-j' (above)
-  - 'demand-margin:j' (above)
-  - 'demand:j' (above)
-  - 'f:' (above)
-  - 'supply-margin:i' (above)
-  - 'supply:i' (above)
-  - 'x:i-j' (above)
-  - 'z:' (above)
